Move stream IO APIs to apis crate
