+++ crates/apis/Cargo.toml
@@ -7,6 +7,7 @@ license.workspace = true
 
 [features]
 console = []
+stream_io = []
 
 [dependencies]
 anyhow = { workspace = true }
