+++ crates/core/Cargo.toml
@@ -16,7 +16,7 @@ crate-type = ["cdylib"]
 [dependencies]
 anyhow = { workspace = true }
 javy = { path = "../javy" }
-javy-apis = { path = "../apis", features = ["console"] }
+javy-apis = { path = "../apis", features = ["console", "stream_io"] }
 once_cell = { workspace = true }
 
 [features]
