+++ crates/core/src/globals.rs
@@ -2,15 +2,11 @@ use anyhow::anyhow;
 use javy::quickjs::{CallbackArg, JSContextRef, JSError, JSValue};
 use javy::Runtime;
 use std::borrow::Cow;
-use std::io::{Read, Write};
 use std::str;
 
 pub fn inject_javy_globals(runtime: &Runtime) -> anyhow::Result<()> {
     let context = runtime.context();
     let global = context.global_object()?;
-
-    let javy_object = context.object_value()?;
-    global.set_property("Javy", javy_object)?;
 
     global.set_property(
         "__javy_decodeUtf8BufferToString",
@@ -21,57 +17,10 @@ pub fn inject_javy_globals(runtime: &Runtime) -> anyhow::Result<()> {
         context.wrap_callback(encode_js_string_to_utf8_buffer())?,
     )?;
 
-    global.set_property(
-        "__javy_io_writeSync",
-        context.wrap_callback(|_, _this_arg, args| {
-            let [fd, data, offset, length, ..] = args else {
-                anyhow::bail!("Invalid number of parameters");
-            };
-
-            let mut fd: Box<dyn Write> = match fd.try_into()? {
-                1 => Box::new(std::io::stdout()),
-                2 => Box::new(std::io::stderr()),
-                _ => anyhow::bail!("Only stdout and stderr are supported"),
-            };
-            let data: Vec<u8> = data.try_into()?;
-            let offset: usize = offset.try_into()?;
-            let length: usize = length.try_into()?;
-            let data = &data[offset..(offset + length)];
-            let n = fd.write(data)?;
-            fd.flush()?;
-            Ok(n.into())
-        })?,
-    )?;
-
-    global.set_property(
-        "__javy_io_readSync",
-        context.wrap_callback(|_, _this_arg, args| {
-            let [fd, data, offset, length, ..] = args else {
-                anyhow::bail!("Invalid number of parameters");
-            };
-            let mut fd: Box<dyn Read> = match fd.try_into()? {
-                0 => Box::new(std::io::stdin()),
-                _ => anyhow::bail!("Only stdin is supported"),
-            };
-            let offset: usize = offset.try_into()?;
-            let length: usize = length.try_into()?;
-            let data = unsafe { data.inner_value() };
-            if !data.is_array_buffer() {
-                anyhow::bail!("Data needs to be an ArrayBuffer");
-            }
-            let data = data.as_bytes_mut()?;
-            let data = &mut data[offset..(offset + length)];
-            let n = fd.read(data)?;
-            Ok(n.into())
-        })?,
-    )?;
-
     context.eval_global(
         "text-encoding.js",
         include_str!("../prelude/text-encoding.js"),
     )?;
-
-    context.eval_global("io.js", include_str!("../prelude/io.js"))?;
 
     Ok(())
 }
